Commenting is done in util, teleop, telefunctions,and globalfunctions
