+++ TeamCode/src/main/java/globalfunctions/Optimizer.java
@@ -5,15 +5,20 @@ import com.qualcomm.robotcore.util.ElapsedTime;
 import java.util.ArrayList;
 
 public class Optimizer {
+    //List of delta times between updates
     public ArrayList<Double> times = new ArrayList<>();
+    //Timer to record times
     public ElapsedTime timer = new ElapsedTime();
+    //Last time
     public double lastTime = 0;
+    //Change in time
     public double deltaTime = 0;
+    //Average change in time
     public double avgDeltaTime = 0;
-
+    //Should show
     public boolean show = false;
 
-
+    //Update the list with delta times
     public void update(){
         if(!show) {
             deltaTime = timer.seconds() - lastTime;
@@ -22,11 +27,16 @@ public class Optimizer {
         }
     }
 
+    //Shows the avg delta time
     public void show(){
         show = true;
         avgDeltaTime = calcAvg(times);
     }
-
+    //Gets the average refresh rate
+    public double getRefreshRate(){
+        return 1/avgDeltaTime; //in htz
+    }
+    //Resets the timer
     public void reset(){
         lastTime = timer.seconds();
         deltaTime = 0;
@@ -34,7 +44,7 @@ public class Optimizer {
         show = false;
         times = new ArrayList<>();
     }
-
+    //Calculates the average of an arraylist of doubles
     public static double calcAvg(ArrayList<Double> in){
         double total = 0;
         for(double i:in){
@@ -43,7 +53,7 @@ public class Optimizer {
 
         return total;
     }
-
+    //Calculates the maximuim of an arraylist of doubles
     public static double max(ArrayList<Double> in){
         double max = 0;
         for(double i:in){
@@ -53,11 +63,11 @@ public class Optimizer {
         }
         return max;
     }
-
+    //Checks if the quantity is within a range
     public static boolean inRange(double in, double[] range){
         return in > range[0] && in < range[1];
     }
-
+    //Calculates the weighted average of a list
     public static double weightedAvg(double[] in, double[] weights){
         double sum = 0;
         double wsum = 0;
@@ -67,7 +77,7 @@ public class Optimizer {
         }
         return sum/wsum;
     }
-
+    //Optimizes heading to [-180,180] range
     public static double optimizeHeading(double heading){
         if(!inRange(heading, new double[]{-180,180})){
             double heading2 = (Math.abs(heading)%360);
@@ -81,11 +91,11 @@ public class Optimizer {
             return heading;
         }
     }
-
+    //Optimizes the heading from a pos
     public static double[] optimizePos(double[] in){
         return new double[]{in[0], in[1], optimizeHeading(in[2])};
     }
-
+    //gets an array of doubles from a string
     public static double[] fromString(String string) {
         String[] strings = string.replace("[", "").replace("]", "").split(", ");
         double[] result = new double[strings.length];
