+++ TeamCode/src/main/java/globalfunctions/Sleep.java
@@ -1,9 +1,11 @@
 package globalfunctions;
 
 public class Sleep {
+    //Interface for runnable
     public interface InterruptedExceptionRunnable {
         void run() throws InterruptedException;
     }
+    //Method to sleep for certain time
     public static void trySleep(InterruptedExceptionRunnable runnable) {
         try { runnable.run(); } catch (InterruptedException ignored) { }
     }
