+++ TeamCode/src/main/java/teleop/OdometryOp.java
@@ -3,7 +3,6 @@ package teleop;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-import com.qualcomm.robotcore.hardware.Gamepad;
 
 import global.TerraBot;
 import globalfunctions.Optimizer;
